order-parent 将金额修改为BigDecimal

+++ cqrs-bookstore/order-context-parent/eventbus-contract/src/main/java/se/citerus/cqrs/bookstore/ordercontext/order/OrderLine.java
@@ -2,13 +2,15 @@ package se.citerus.cqrs.bookstore.ordercontext.order;
 
 import se.citerus.cqrs.bookstore.TransportObject;
 
+import java.math.BigDecimal;
+
 public class OrderLine extends TransportObject {
     public final ProductId productId;
     public final String title;
     public final int quantity;
-    public final long unitPrice;
+    public final BigDecimal unitPrice;
 
-    public OrderLine(ProductId productId, String title, int quantity, long unitPrice) {
+    public OrderLine(ProductId productId, String title, int quantity, BigDecimal unitPrice) {
         this.productId = productId;
         this.title = title;
         this.quantity = quantity;
