+++ cqrs-bookstore/order-context-parent/eventbus-contract/src/main/java/se/citerus/cqrs/bookstore/ordercontext/publishercontract/event/PurchaseRegisteredEvent.java
@@ -4,13 +4,16 @@ import se.citerus.cqrs.bookstore.event.DomainEvent;
 import se.citerus.cqrs.bookstore.ordercontext.order.ProductId;
 import se.citerus.cqrs.bookstore.ordercontext.publishercontract.PublisherContractId;
 
+import java.math.BigDecimal;
+
 public class PurchaseRegisteredEvent extends DomainEvent<PublisherContractId> {
     public final ProductId productId;
-    public final long purchaseAmount;
-    public final long feeAmount;
-    public final long accumulatedFee;
+    public final BigDecimal purchaseAmount;
+    public final BigDecimal feeAmount;
+    public final BigDecimal accumulatedFee;
 
-    public PurchaseRegisteredEvent(PublisherContractId aggregateId, int version, long timestamp, ProductId productId, long purchaseAmount, long feeAmount, long accumulatedFee) {
+    public PurchaseRegisteredEvent(PublisherContractId aggregateId, int version, long timestamp, ProductId productId,
+                                   BigDecimal purchaseAmount, BigDecimal feeAmount, BigDecimal accumulatedFee) {
         super(aggregateId, version, timestamp);
         this.productId = productId;
         this.purchaseAmount = purchaseAmount;
