+++ cqrs-bookstore/order-context-parent/order-command/src/main/java/se/citerus/cqrs/bookstore/ordercontext/api/LineItemDto.java
@@ -6,6 +6,7 @@ import javax.validation.constraints.Min;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Pattern;
+import java.math.BigDecimal;
 
 public class LineItemDto {
 
@@ -17,11 +18,11 @@ public class LineItemDto {
     public String title;
 
     @Min(1)
-    public long price;
+    public BigDecimal price;
 
     @Min(1)
     public int quantity;
 
     @Min(1)
-    public long totalPrice;
+    public BigDecimal totalPrice;
 }
