+++ cqrs-bookstore/order-context-parent/order-query/src/main/java/se/citerus/cqrs/bookstore/ordercontext/query/orderlist/OrderLineProjection.java
@@ -2,9 +2,11 @@ package se.citerus.cqrs.bookstore.ordercontext.query.orderlist;
 
 import se.citerus.cqrs.bookstore.ordercontext.order.ProductId;
 
+import java.math.BigDecimal;
+
 public class OrderLineProjection {
     public ProductId productId;
     public String title;
     public int quantity;
-    public long unitPrice;
+    public BigDecimal unitPrice;
 }
